Move search functions: showAvailableVideos.js -> search.js
