+++ client/scripts/search.js
@@ -1,3 +1,99 @@
+import * as basic from "../scripts/basics.js";
+import * as showAvailableVideos from "../scripts/showAvailableVideos.js";
+
+export function searchBarContainer() {
+    let searchBarContainer = document.getElementById("searchBarContainer");
+    if (!searchBarContainer) {
+        searchBarContainer = basic.createSection(basic.websiteContentContainer(), "section", "searchBarContainer", "searchBarContainer"); 
+    }    
+    return searchBarContainer;
+}
+
+// reset search bar value
+export function resetSearchBarValue() {
+    if (document.getElementById("searchBar")) {
+        document.getElementById("searchBar").value = ""; 
+    }
+}
+
+// find video by filtering trough each available video by textinput
+export function searchBar(){
+    // create search input
+    const searchBar = basic.inputType(searchBarContainer(), "text", "searchBar", "searchBar", true);
+    searchBar.name = "searchBar";
+    searchBar.placeholder="Type to search";
+    // filters trough video data by name at every key press
+    searchBar.addEventListener("keyup", (e) => { 
+        const searchString = e.target.value;
+        searchBarKeyUp(searchString);
+    });
+    return "searchBar";
+}
+  
+// filters trough video data by searchString input
+export function searchBarKeyUp(searchString) { 
+    if (typeof searchString == "string") {
+        const savedVideosThumbnailContainer = document.getElementById("savedVideosThumbnailContainer");
+        // check from searchableVideoDataArray if any video data title matches input string
+        const filteredsearchableVideoData = getSearchableVideoDataArray().filter((video) => {
+            return (
+                video.info.title.toLowerCase().includes(searchString.toLowerCase())
+            );
+        }); 
+        // clear savedVideosThumbnailContainer
+        savedVideosThumbnailContainer.innerHTML = ""; 
+        // check if inputed key phrase available data is avaiable or not to either display data or state the problem
+        if (filteredsearchableVideoData.length == 0) {
+            noAvailableOrSearchableVideoMessage();
+        } else {
+            showAvailableVideos.removeNoAvailableVideosDetails();
+            removeNoSearchableVideoData();
+            // display filterd details to client
+            filteredsearchableVideoData.forEach(function(data) {   
+                if (data.info.id.includes("folder-")) {
+                    showAvailableVideos.showFolderDetails(savedVideosThumbnailContainer, data.info.id, data);
+                } else { 
+                    showAvailableVideos.showDetails(savedVideosThumbnailContainer, data.info.id, data);
+                } 
+            });
+            return "Display filterd avaiable video data";
+        } 
+    } else {
+        return "searchString not string";
+    }
+}
+
+
+// display noSearchableVideoData no if exits
+export function noSearchableVideoData() {
+    if (!document.getElementById("noSearchableVideoData")) {
+        const noSearchableVideoData = basic.createSection(basic.websiteContentContainer(), "section", "noAvailableVideosContainer", "noSearchableVideoData");
+        basic.createSection(noSearchableVideoData, "h1", "noAvailableVideosHeader", undefined,  "No results found: Try different keywords");
+    }
+}
+
+// remove noSearchableVideoData if exits
+export function removeNoSearchableVideoData() {
+    if (document.getElementById("noSearchableVideoData")) {
+        document.getElementById("noSearchableVideoData").remove();
+    }
+}
+  
+// if savedVideosThumbnailContainer is empty, display either noAvailableVideosDetails or noSearchableVideoData depending on the senario 
+export function noAvailableOrSearchableVideoMessage() {
+    if (document.getElementById("savedVideosThumbnailContainer")) {
+        if (document.getElementById("savedVideosThumbnailContainer").childElementCount == 0) {  
+            if(getSearchableVideoDataArray().length == 0){ 
+                showAvailableVideos.noAvailableVideosDetails();
+                return "no avaiable video data";
+            } else {
+                noSearchableVideoData();
+                return "key phrase unavailable";
+            }
+        }
+    } 
+}
+  
 export let searchableVideoDataArray = []; 
 
 // push data to SearchableVideoDataArray
@@ -22,7 +118,6 @@ export function deleteIDFromSearchableVideoDataArray(id) {
     return "updated SearchableVideoDataArray";
 }
   
-
 // return SearchableVideoDataArray to its inital state
 export function resetSearchableVideoDataArray() {
     searchableVideoDataArray = [];
@@ -88,4 +183,3 @@ export function searchableVideoDataArray_move_after(from_id, to_id) {
         } 
     }
 }
-
